Fixing dcos-history service to get data
